requirements.txt
selenium==4.17.2
pandas==2.2.0
Flask==3.0.1
SQLAlchemy==2.0.25
tabula-py==2.9.0
inflect==7.0.0